refactor : mocks/api/data -> mocks/mockData 로 mockingData 파일 폴더 이름 변경및 위치 변경
