+++ src/utils/mocks/mockData/auctionPost/auctionComment.ts
@@ -1,4 +1,4 @@
-import { CommentType } from "../../types";
+import { CommentType } from "../../api/types";
 
 export const commentData: CommentType[] = [
   {
