+++ src/utils/mocks/mockData/auctionPost/auctionDetail.ts
